third_party/blink/web_tests/fast/events/clipboard-dataTransferItemList-remove-expected.txt

Checks that DataTransferItemList.remove() is working

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

* copy event
PASS dataTansferItemList.length is 4
PASS dataTansferItemList[0].type is "text/plain"
PASS dataTansferItemList[1].type is "text/uri-list"
PASS dataTansferItemList[2].type is "text/html"
PASS dataTansferItemList[3].type is "custom-data"
PASS dataTansferItemList.remove() threw exception TypeError: Failed to execute 'remove' on 'DataTransferItemList': 1 argument required, but only 0 present..
PASS dataTansferItemList.remove(-1) did not throw exception.
PASS dataTansferItemList.length is 4
PASS dataTansferItemList.remove(999) did not throw exception.
PASS dataTansferItemList.length is 4
PASS dataTansferItemList.remove(0) did not throw exception.
PASS dataTansferItemList.length is 3
PASS dataTansferItemList[0].type is "text/uri-list"
PASS dataTansferItemList[1].type is "text/html"
PASS dataTansferItemList.remove(-4294967295) did not throw exception.
PASS dataTansferItemList.length is 2
PASS dataTansferItemList[1].type is "custom-data"
* paste event
PASS dataTansferItemList.length is 2
PASS dataTansferItemList.remove(0) threw exception InvalidStateError: Failed to execute 'remove' on 'DataTransferItemList': The list is not writable..
PASS dataTansferItemList.length is 2
PASS successfullyParsed is true

TEST COMPLETE
